feat: add path to test server
allow, optionally, the ability to specify a path in a handler. When handling a response, the handler will first look for a handler that matches the path, if it does not find one it will use a handler without a name (effectively matching any path). handlers set with the same path function the same way handlers have functioned, looping through available handlers by index